backend/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# LangChain & Gemini
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
